Show correct track name in scratchpad
