+++ test/system/components/mentoring/discussion_test.rb
@@ -306,81 +306,6 @@ module Components
         refute_text "Edit"
       end
 
-      test "mentor saves scratchpad page" do
-        mentor = create :user, handle: "author"
-        solution = create :concept_solution
-        discussion = create :mentor_discussion, solution: solution, mentor: mentor
-        create :iteration, solution: solution
-
-        use_capybara_host do
-          sign_in!(mentor)
-          visit test_components_mentoring_discussion_path(discussion_id: discussion.id)
-          click_on "Scratchpad"
-          fill_in_editor "# Hello"
-          assert_text "Unsaved"
-          click_on "Save"
-          assert_no_text "Unsaved"
-        end
-      end
-
-      test "mentor sees scratchpad page" do
-        mentor = create :user, handle: "author"
-        exercise = create :concept_exercise
-        solution = create :concept_solution, exercise: exercise
-        discussion = create :mentor_discussion, solution: solution, mentor: mentor
-        submission = create :submission, solution: solution
-        create :iteration, solution: solution, submission: submission
-        create :scratchpad_page, content_markdown: "# Some notes", author: mentor, about: exercise
-
-        use_capybara_host do
-          sign_in!(mentor)
-          visit test_components_mentoring_discussion_path(discussion_id: discussion.id)
-          click_on "Scratchpad"
-
-          assert_editor_text "# Some notes"
-          assert_no_text "Unsaved"
-        end
-      end
-
-      test "mentor updates scratchpad page" do
-        mentor = create :user, handle: "author"
-        exercise = create :concept_exercise
-        solution = create :concept_solution, exercise: exercise
-        discussion = create :mentor_discussion, solution: solution, mentor: mentor
-        submission = create :submission, solution: solution
-        create :iteration, solution: solution, submission: submission
-        create :scratchpad_page, content_markdown: "# Some notes", author: mentor, about: exercise
-
-        use_capybara_host do
-          sign_in!(mentor)
-          visit test_components_mentoring_discussion_path(discussion_id: discussion.id)
-          click_on "Scratchpad"
-          fill_in_editor "# Hello"
-          assert_text "Unsaved"
-          click_on "Save"
-          assert_no_text "Unsaved"
-        end
-      end
-
-      test "mentor reverts scratchpad page" do
-        mentor = create :user, handle: "author"
-        exercise = create :concept_exercise
-        solution = create :concept_solution, exercise: exercise
-        discussion = create :mentor_discussion, solution: solution, mentor: mentor
-        submission = create :submission, solution: solution
-        create :iteration, solution: solution, submission: submission
-        create :scratchpad_page, content_markdown: "# Some notes", author: mentor, about: exercise
-
-        use_capybara_host do
-          sign_in!(mentor)
-          visit test_components_mentoring_discussion_path(discussion_id: discussion.id)
-          click_on "Scratchpad"
-          fill_in_editor "# Hello"
-          click_on "Revert"
-          assert_editor_text "# Some notes"
-        end
-      end
-
       test "mentor marks discussion as nothing to do" do
         mentor = create :user, handle: "author"
         exercise = create :concept_exercise
